repair tiem

+++ src/libs/MenuLogic.h
@@ -3,6 +3,6 @@
 int ShowMenu();
 void MenuLogic(int, int);
 int SwitchMode(int, int);
-float Start(int);
+int Start(int);
 int ChooseDifficulty();
 void Results();
